Safari-611.1.10/LayoutTests/editing/pasteboard/data-transfer-is-unique-for-dragenter-and-dragleave-expected.txt

PASS ev.__proto__ is DragEvent.prototype
PASS dragstart for source had an unique dataTransfer object
PASS ev.__proto__ is DragEvent.prototype
PASS dragenter for passover had an unique dataTransfer object
PASS ev.__proto__ is DragEvent.prototype
PASS dragenter for destination had an unique dataTransfer object
PASS ev.__proto__ is DragEvent.prototype
PASS dragleave for passover had an unique dataTransfer object
PASS ev.__proto__ is DragEvent.prototype
PASS dragover for destination had an unique dataTransfer object
PASS ev.__proto__ is DragEvent.prototype
PASS drop for destination had an unique dataTransfer object
PASS successfullyParsed is true

TEST COMPLETE
